Local counter: Don't re-allocate maps in Go 1.21+
